phase-5/services/notification-service/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = "Service to handle notifications"
authors = ["Developer <developer@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
kafka-python = "^2.0.2"
dapr-ext-fastapi = "^1.9.0"
web-push = "^3.0.1"
sendgrid = "^6.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"